% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_dir_exists}
\alias{assert_dir_exists}
\title{Assert a directory exists on disk}
\usage{
assert_dir_exists(x)
}
\arguments{
\item{x}{[chr] A directory path}
}
\value{
[none] stop if assertion fails
}
\description{
Assert a directory exists on disk
}
\seealso{
Other assertions: 
\code{\link{assert_named_list}()},
\code{\link{assert_scalar}()},
\code{\link{assert_scalar_not_empty}()},
\code{\link{assert_type}()},
\code{\link{assert_x_in_y}()}
}
\concept{assertions}
\keyword{internal}
