% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_x_in_y}
\alias{assert_x_in_y}
\title{Assert all elements of x are in y}
\usage{
assert_x_in_y(x, y)
}
\arguments{
\item{x}{[vector] some vector}

\item{y}{[vector] some vector}
}
\value{
[none] stop if any elements of x are not in y
}
\description{
Assert all elements of x are in y
}
\seealso{
Other assertions: 
\code{\link{assert_dir_exists}()},
\code{\link{assert_named_list}()},
\code{\link{assert_scalar}()},
\code{\link{assert_scalar_not_empty}()},
\code{\link{assert_type}()}
}
\concept{assertions}
\keyword{internal}
