% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_path.R
\name{BlankIfNA}
\alias{BlankIfNA}
\title{Replace NAs by a user-defined string}
\usage{
BlankIfNA(x, blank = "")
}
\arguments{
\item{x}{[vector] whose NAs should be overwritten with 0s.}

\item{blank}{[chr] the value to replace NAs with.}
}
\value{
[vector] edited vector x
}
\description{
replaces - DescTools::BlankIfNA()
}
\details{
Replace NAs in a numeric vector x with 0. This function has the same
logic as the zeroifnull function in SQL. NAIfZero() does replace zeros
with NA. BlankIfNA() and NAIfBlank() do the same, but for character
vectors.
}
\examples{
BlankIfNA(c(1, 2, NA, 4)) # "1" "2" "" "4"
}
