% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvignon.R
\name{calcAvignon.SI}
\alias{calcAvignon.SI}
\title{Calculate Avignon SI}
\usage{
calcAvignon.SI(
  insulin.fasting,
  glucose.fasting,
  insulin.2h,
  glucose.2h,
  weight
)
}
\arguments{
\item{weight}{in kg}

\item{fasting.insulin}{in μU/mL}

\item{fasting.glucose}{in mass units (mg/dL)}

\item{2h.insulin}{in μU/mL}

\item{2h.glucose}{in mass units (mg/dL)}
}
\value{
Avignon SI.b, SI.2h and SI.m as numeric value
}
\description{
This function calculates Avignon insulin sensitivity index from 2 hour oral glucose tolerance test
Reference: Avignon A, et al. International Journal of Obesity 1999; 23:512-7
DOI: 10.1038/sj.ijo.0800864
}
\examples{
calcAvignon.SI(15.0,100.0,30,150,80)

glucose.fasting.vals <- seq(75,175,by=25)
insulin.fasting.vals <- seq(2,20, by=1)
glucose.2h.vals <- seq(75,175,by=25)
insulin.2h.vals <- seq(2,20, by=1)
weight.vals <- seq(50,90,by=10)
test.dat <- expand.grid(insulin.fasting=insulin.vals,glucose.fasting=glucose.vals,
                        insulin.2h=insulin.2h.vals,glucose.2h=glucose.2h.vals,weight=weight.vals)

test.dat <- cbind(test.dat,calcAvignon.SI(insulin.fasting=test.dat$insulin.fasting,
                                          glucose.fasting=test.dat$glucose.fasting,
                                          insulin.2h=test.dat$insulin.2h,
                                          glucose.2h=test.dat$glucose.2h,
                                          weight=test.dat$weight))
}
\keyword{Avignon}
\keyword{SI.b}
