% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHOMA.beta.R
\name{calcHOMA.beta}
\alias{calcHOMA.beta}
\title{Calculate HOMA-β}
\usage{
calcHOMA.beta(insulin, glucose)
}
\arguments{
\item{insulin}{in μU/mL}

\item{glucose}{in mass units (mg/dL)}
}
\value{
HOMA-β cell function as percentage value
}
\description{
This function calculates Homeostatic Model Assessment β-cell function
Reference: Matthews DR, et al. Diabetologia 1995;28:412-9
DOI: 10.1007%2FBF00280883
}
\examples{
calcHOMA.beta(15.0,100.0)

glucose.vals <- seq(75,175,by=25)
insulin.vals <- seq(2,20, by=1)
test.dat <- expand.grid(insulin=insulin.vals,glucose=glucose.vals)

test.dat$HOMA.beta <- calcHOMA.beta(insulin=test.dat$insulin,glucose=test.dat$glucose)
}
\keyword{HOMA-β}
