% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMatsuda.ISI.R
\name{calcMatsuda.ISI}
\alias{calcMatsuda.ISI}
\title{Calculate Matsuda ISI}
\usage{
calcMatsuda.ISI(insulin.fasting, glucose.fasting, insulin.mean, glucose.mean)
}
\arguments{
\item{insulin.fasting}{in μU/mL}

\item{glucose.fasting}{in mass units (mg/dL)}

\item{insulin.mean}{in μU/mL}

\item{glucose.mean}{in mass units (mg/dL)}
}
\value{
Matsuda ISI as numeric value
}
\description{
This function calculates Matsuda insulin sensitivity index from 2 hour oral glucose tolerance test
Reference: Matsuda M and DeFronzo RA. Diabetes Care 1999; 22:1462-70
DOI: 10.2337/diacare.22.9.1462
}
\examples{
calcMatsuda.ISI(15.0,100.0,30,150)

glucose.fasting.vals <- seq(75,175,by=25)
insulin.fasting.vals <- seq(2,20, by=1)
glucose.mean.vals <- seq(75,175,by=25)
insulin.mean.vals <- seq(2,20, by=1)
test.dat <- expand.grid(insulin.fasting=insulin.vals,glucose.fasting=glucose.vals,
                        insulin.mean=insulin.mean.vals,glucose.mean=glucose.mean.vals)

test.dat$Matsuda.ISI <- calcMatsuda.ISI(insulin.fasting=test.dat$insulin.fasting,
                                        glucose.fasting=test.dat$fastingglucose,
                                        insulin.mean=test.dat$insulin.mean,
                                        glucose.mean=test.dat$glucose.mean)
}
\keyword{ISI}
\keyword{Matsuda}
