% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCEcalc.R
\name{calcPCE}
\alias{calcPCE}
\title{Calculate Pooled Cohorts Equations risk scores}
\usage{
calcPCE(age, race, sex, tc, hdl, sbp, bptx, smoking, dm)
}
\arguments{
\item{age}{age in years}

\item{race}{0=white/1=black}

\item{sex}{0=female/1=male}

\item{tc}{total cholesterol in mg/dl}

\item{hdl}{HDL cholesterol in mg/dl}

\item{sbp}{Systolic blood pressure in mmHg}

\item{bptx}{treatment for blood pressure 0=FALSE/1=TRUE}

\item{smoking}{active smoking 0=FALSE/1=TRUE}

\item{dm}{diabetes mellitus 0=FALSE/1=TRUE}
}
\value{
PCE risk as numeric from 0 to 1
}
\description{
This function calculates Pooled Cohorts Equations risk scores per recent ACC/AHA recommendations.
Reference: Goff DC, et al. Circulation 2014;19:S49-S73
DOI: 10.1161/01.cir.0000437741.48606.98
}
\examples{
calcPCE(55,0,0,213,50,120,0,0,0)
calcPCE(55,1,0,213,50,120,0,0,0)
calcPCE(55,0,1,213,50,120,0,0,0)
calcPCE(55,1,1,213,50,120,0,0,0)


test.dat <- data.frame(age=rep(56:65,4),race=c(rep(0,10),rep(1,10),rep(0,10),rep(1,10)),
                       sex=c(rep(0,20),rep(1,20)),tc=rep(136:145,4),hdl=rep(46:55,4),sbp=rep(126:135,4),
                       bptx=rep(c(rep(0,5),rep(1,5)),4),smoking=rep(c(rep(0,5),rep(1,5)),4),
                       dm=rep(c(rep(0,5),rep(1,5)),4))

test.dat$PCE <- calcPCE(test.dat$age,test.dat$race,test.dat$sex,test.dat$tc,test.dat$hdl,test.dat$sbp,
                        test.dat$bptx,test.dat$smoking,test.dat$dm)
}
\keyword{PCE}
