% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPREVENT.ASCVD.R
\name{calcPREVENT.ASCVD}
\alias{calcPREVENT.ASCVD}
\title{Calculate PREVENT 10 year ASCVD risk scores}
\usage{
calcPREVENT.ASCVD(
  age,
  sex,
  tc,
  hdl,
  sbp,
  bptx,
  smoking,
  dm,
  statin,
  egfr,
  female = c(0, "f", "female"),
  male = c(1, "m", "male"),
  bptx.true = c(1, "t", "true", "y", "yes"),
  bptx.false = c(0, "f", "false", "n", "no"),
  smoking.true = c(1, "t", "true", "y", "yes", "active"),
  smoking.false = c(0, "f", "false", "former", "non-smoker", "nonsmoker"),
  dm.true = c(1, "t", "true", "y", "yes"),
  dm.false = c(0, "f", "false", "n", "no"),
  statin.true = c(1, "t", "true", "y", "yes"),
  statin.false = c(0, "f", "false", "n", "no")
)
}
\arguments{
\item{age}{age in years}

\item{sex}{0=female/1=male}

\item{tc}{total cholesterol in mg/dl}

\item{hdl}{HDL cholesterol in mg/dl}

\item{sbp}{Systolic blood pressure in mmHg}

\item{bptx}{treatment for blood pressure 0=FALSE/1=TRUE}

\item{smoking}{active smoking 0=FALSE/1=TRUE}

\item{dm}{diabetes mellitus 0=FALSE/1=TRUE}

\item{statin}{use of statin medication 0=FALSE/1=TRUE}

\item{egfr}{estimated glomerular filtration rate in ml/min/1.73m^2}

\item{female}{values which indicate female in the parameter sex, case-insensitive (default=0,"f","female")}

\item{male}{values which indicate male in the parameter sex, case-insensitive (default=1,"m","male")}

\item{bptx.true}{values which indicate BP med use, case-insensitive (default=1,"t","true","y","yes")}

\item{bptx.false}{values which indicate no BP med use, case-insensitive (default=0,"f","false","n","no")}

\item{smoking.true}{values which indicate active smoking, case-insensitive (default=1,"t","true","y","yes","active")}

\item{smoking.false}{values which indicate no active smoking, case-insensitive (default=0,"f","false","n","no","former","non-smoker","nonsmoker")}

\item{dm.true}{values which indicate diabetes, case-insensitive (default=1,"t","true","y","yes")}

\item{dm.false}{values which indicate no diabetes, case-insensitive (default=0,"f","false","n","no")}

\item{statin.true}{values which indicate statin use, case-insensitive (default=1,"t","true","y","yes")}

\item{statin.false}{values which indicate no statin use, case-insensitive (default=0,"f","false","n","no")}
}
\value{
PREVENT 10 year ASCVD risk as numeric from 0 to 1
}
\description{
This function calculates PREVENT 10 year ASCVD risk scores per recent AHA recommendations.
Reference: Khan SS, et al. Circulation 2023 online ahead of print
DOI: https://doi.org/10.1161/CIRCULATIONAHA.123.067626
}
\examples{
calcPREVENT.ASCVD(50,0,200,45,160,1,0,1,0,90)
calcPREVENT.ASCVD(55,1,213,50,120,0,0,0,0,60)
calcPREVENT.ASCVD(55,0,213,50,120,0,0,0,0,60)
calcPREVENT.ASCVD(55,1,213,50,120,0,0,0,0,60)


test.dat <- data.frame(age=rep(56:65,4),sex=c(rep(0,20),rep(1,20)),tc=rep(136:145,4),hdl=rep(46:55,4),
                       sbp=rep(126:135,4),bptx=rep(c(rep(0,5),rep(1,5)),4),
                       smoking=rep(c(rep(0,5),rep("active",5)),4),dm=rep(c(rep("No",5),rep(1,5)),4),
                       statin=c(rep(0,10),rep(1,10),rep(0,10),rep(1,10)),
                       egfr=c(rep(40,10),rep(90,10),rep(40,10),rep(90,10)))

test.dat$PREVENT.ASCVD <- calcPREVENT.ASCVD(test.dat$age,test.dat$sex,test.dat$tc,test.dat$hdl,test.dat$sbp,
                                            test.dat$bptx,test.dat$smoking,test.dat$dm,test.dat$statin,test.dat$egfr)
}
\keyword{PREVENT}
