% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQUICKI.R
\name{calcQUICKI}
\alias{calcQUICKI}
\title{Calculate QUICKI}
\usage{
calcQUICKI(insulin, glucose)
}
\arguments{
\item{insulin}{in μU/mL}

\item{glucose}{in mass units (mg/dL)}
}
\value{
QUICKI as numeric value
}
\description{
This function calculates quantitative insulin sensitivity check index
Reference: Katz A, et al. Journal of Clinical Endocrinology & Metabolism 2000;85:2402-10
DOI: 10.1210/jcem.85.7.6661
}
\examples{
calcQUICKI(15.0,100.0)

glucose.vals <- seq(75,175,by=25)
insulin.vals <- seq(2,20, by=1)
test.dat <- expand.grid(insulin=insulin.vals,glucose=glucose.vals)

test.dat$QUICKI <- calcQUICKI(insulin=test.dat$insulin,glucose=test.dat$glucose)
}
\keyword{QUICKI}
