% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_tree.R
\name{dir_tree}
\alias{dir_tree}
\title{Print a directory tree to stdout}
\usage{
dir_tree(path = ".", level = Inf, prefix = "")
}
\arguments{
\item{path}{[chr] The path to the directory to print}

\item{level}{[int] The maximum depth to print}

\item{prefix}{[chr] The prefix to add to each line}
}
\description{
Print a directory tree to stdout
}
