% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlink_tool.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{WARNING: DO NOT touch '.__enclos_env__' unless you want the tool to break}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
NOTES:
  DEBUGGING:
    - The 'red dot' stop point will not work with R6 objects
    - Use `debug(my_fun)/undebug(my_fun)` or place a `browser()` in the code
    - See https://r6.r-lib.org/articles/Debugging.html
}
\details{
LIBRARIES
    - Depends on vaccines/vaccination_pipeline_functions/validations.R for assert_data_schema
    - Relies on libraries declared above, but calls on all functions by package namespace for clarity and robustness
}
\keyword{datasets}
