% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_version_name.R
\name{find_latest_output_dir}
\alias{find_latest_output_dir}
\title{Find the latest output directory with format YYYY_MM_DD.VV}
\usage{
find_latest_output_dir(root)
}
\arguments{
\item{root}{[chr] path to root of output results}
}
\value{
[chr] path to latest output directory
}
\description{
Used only for signaling/messaging
}
\keyword{internal}
