% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_n_cores.R
\name{find_n_cores}
\alias{find_n_cores}
\title{Cross platform helper to find number of cores}
\usage{
find_n_cores()
}
\value{
[int]
}
\description{
Cross platform helper to find number of cores
}
\keyword{internal}
