% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatpval.R
\name{format.pval}
\alias{format.pval}
\title{Format p-value for pretty printing}
\usage{
\method{format}{pval}(p, include.p = TRUE, na.string = "NA")
}
\arguments{
\item{p}{vector of p-values}

\item{include.p}{if TRUE then the characters "p=" or "p<" are prepended (default:TRUE)}

\item{na.string}{characters to output if p-value is NA (default: "NA")}
}
\description{
Format p-value for pretty printing
}
\examples{
p <- c(0.5,0.499999,0.004,0.0003,1e-10,NA)
format.pval(p)
}
