% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{Get labels from a dataset}
\usage{
get_labels(ds)
}
\arguments{
\item{ds}{dataset}
}
\value{
named vector with values of the column labels and names of column names
}
\description{
This function pulls out all of the labels from a dataset
}
\examples{
get_labels(mtcars)
}
