% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_version_name.R
\name{get_new_version_name}
\alias{get_new_version_name}
\title{Increment a new output folder version as "YYYY_MM_DD.VV"}
\usage{
get_new_version_name(root, date = "today")
}
\arguments{
\item{root}{[chr] path to root of output results}

\item{date}{[chr] character date in form of "YYYY_MM_DD" or "today". "today" will be interpreted as today's date.}
}
\value{
[chr] new output version of the form "YYYY_MM_DD.VV"
}
\description{
Return on the date-version, not the full path.  Does not create a folder.
}
\examples{
get_new_version_name(root = tempdir(), date = "today") # expect "YYYY_MM_DD.01"
}
