% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB_ID_from_name.R
\name{HMDB_ID_from_name}
\alias{HMDB_ID_from_name}
\title{Query Human Metabolome Database by metabolite name}
\usage{
HMDB_ID_from_name(met.names, max.depth = 25, max.tries = 5)
}
\arguments{
\item{max.depth}{number indicating how many entries to be checked (may be slightly more than this,
depending on entries per page)}

\item{max.tries}{number of times to retry loading a page}

\item{met.name}{string containing metabolite name}
}
\value{
string with canonical HMDB ID
}
\description{
This function searches the HMDB to find the canonical HMDB ID for a given metabolite name. It does look through
the synonym list as well. Importantly, it takes the *first* match among the names returned by the HMDB search engine
and if none, the first match among the synonyms.
}
\examples{
HMDB_ID_from_one_name("valine")
HMDB_ID_from_one_name("ectoine")

test.list <- c("valine-d8","phenylalanine-d8","glycine","alanine","serine","threonine","methionine","glutamate",
"asparagine","glutamine","histidine","arginine","lysine","valine","leucine","isoleucine","phenylalanine",
"tyrosine","tryptophan","proline","hydroxyproline","ornithine","citrulline","taurine","GABA",
"dimethylglycine","ADMA","SDMA","NMMA","2-aminoisobutyric acid","kynurenic acid","1-methylhistamine",
"N-carbamoyl-beta-alanine","thiamine","niacinamide","betaine","PE(P-36:4)/PE(O-36:5)","PE(P-36:2)/PE(O-36:3)",
"PE(P-38:6)/PE(O-38:7)","PE(P-38:4)/PE(O-38:5)","PE(P-40:6)/PE(O-40:7)")

t0 <- Sys.time()
HMDB_ID_from_name(test.list)
Sys.time() -t0

HMDB_ID_from_name("C5 carnitine")
}
\keyword{HMDB}
