% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{is_an_error}
\alias{is_an_error}
\title{Determine if an object is an error}
\usage{
is_an_error(x)
}
\description{
@param x [obj] some R object
}
\details{
@return [lgl] TRUE / FALSE

 @examples
 is_an_error(simpleError("error")) # TRUE
 is_an_error(message("error)) # FALSE
}
