% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_windows_admin.R
\name{is_windows_admin}
\alias{is_windows_admin}
\title{If running on windows, check if the user has admin privileges}
\usage{
is_windows_admin()
}
\value{
[lgl] TRUE if the user in on a windows OS and has admin privileges, FALSE otherwise
}
\description{
If running on windows, check if the user has admin privileges
}
\keyword{internal}
