% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqread.R
\name{mqread}
\alias{mqread}
\title{Fast read wrapper for multiple objects using qs}
\usage{
mqread(fname, ...)
}
\arguments{
\item{fname}{filename}

\item{...}{additional parameters passed to qsave}

\item{objects}{vector containing names of objects to save}
}
\value{
returns result of qsave call from qs module
}
\description{
This reads output of mqsave which wraps objects in a list to allow saving multiple objects using qs
}
\examples{
calcHOMA.IR(15.0,100.0)

test.data <- data.frame(var1=rnorm(1000), var2=rnorm(1000))
mqsave(objects="test.data",fname="test.qs")
rm(test.data)
mqread(fname="test.qs")
}
\keyword{fast}
\keyword{save}
\keyword{serialization}
