% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqsave.R
\name{mqsave}
\alias{mqsave}
\title{Fast save wrapper for multiple objects using qs}
\usage{
mqsave(objects, fname, ...)
}
\arguments{
\item{objects}{vector containing names of objects to save}

\item{fname}{filename}

\item{...}{additional parameters passed to qsave}
}
\value{
returns result of qsave call from qs module
}
\description{
This function wraps multiple objects into a list to allow saving using qs
}
\examples{
calcHOMA.IR(15.0,100.0)

test.data <- data.frame(var1=rnorm(1000), var2=rnorm(1000))
mqsave(objects="test.data",fname="test.qs")
rm(test.data)
mqread(fname="test.qs")
}
\keyword{fast}
\keyword{save}
\keyword{serialization}
