% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_FEV1.R
\name{predict_FEV1}
\alias{predict_FEV1}
\title{Predict FEV1 based on Global Lungs Initiative}
\usage{
predict_FEV1(age, height, race, sex, fev1 = NA)
}
\arguments{
\item{age}{age in years (ages 25-95)}

\item{height}{height in centimeters}

\item{race}{1=African American (AA), 2=North East Asian (NEA), 3=South East Asian (SEA) or 4=other (O)}

\item{sex}{0=female/1=male}

\item{fev1}{measured FEV1 (optional). If provided, Z-score and percent predicted will be computed}
}
\value{
vector of predicted FEV1, coefficient of variation, skewness, lower limit normal,
Z-score (if fev1 is provided), percent predicted FEV1 (if fev1 is provided)
}
\description{
This function computes predicted FEV1 from the ERS Global Lung Initiative function (Eur Resp J 2012 40:1324-43)
}
\examples{

age <- c(4.8, 12.2, 53, 39.1)
height <- c(107, 152, 175, 165)
race <- c("o", "aa", "SEA", "SEA")
sex <- c("M", "M", "M", "F")
fev1 <- c(0.8, 2.405, 2.410, 2.210)

predict_FEV1(age, height, race, sex, fev1)

}
\keyword{FEV1}
