% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_FVC.R
\name{predict_FVC}
\alias{predict_FVC}
\title{Predict FVC based on Global Lungs Initiative}
\usage{
predict_FVC(age, height, race, sex, fvc = NA)
}
\arguments{
\item{age}{age in years (ages 25-95)}

\item{height}{height in centimeters}

\item{race}{1=African American (AA), 2=North East Asian (NEA), 3=South East Asian (SEA) or 4=other (O)}

\item{sex}{0=female/1=male}

\item{fvc}{measured FVC (optional). If provided, Z-score and percent predicted will be computed}
}
\value{
vector of predicted FVC, coefficient of variation, skewness, lower limit normal,
Z-score (if fvc is provided), percent predicted FVC (if fvc is provided)
}
\description{
This function computes predicted FVC from the ERS Global Lung Initiative function (Eur Resp J 2012 40:1324-43)
}
\examples{

age <- c(4.8, 12.2, 53, 39.1)
height <- c(107, 152, 175, 165)
race <- c("o", "aa", "SEA", "SEA")
sex <- c("M", "M", "M", "F")
fvc <- c(0.8, 2.405, 2.410, 2.210)

predict_FVC(age, height, race, sex, fvc)

}
\keyword{FEV1}
