% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_line_breaks.R
\name{split_line_breaks}
\alias{split_line_breaks}
\title{Split a character vector by line breaks}
\usage{
split_line_breaks(string)
}
\arguments{
\item{string}{[chr] A character vector.}
}
\value{
All elements of the character vector are split by `\\n` into new elements.
}
\description{
Split a character vector by line breaks
}
