% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lines.R
\name{split_lines}
\alias{split_lines}
\title{Split a character vector by line breaks}
\usage{
split_lines(x)
}
\arguments{
\item{x}{[chr] A character vector.}
}
\value{
All elements of the character vector are split by `\\n` into lines.
}
\description{
Split a character vector by line breaks
}
\examples{
split_lines(c('a', 'b\nc'))
}
