% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_path.R
\name{split_path}
\alias{split_path}
\title{Split Path In Drive, Path, Filename}
\usage{
split_path(path, last.is.file = NULL)
}
\arguments{
\item{path}{[chr] a path}

\item{last.is.file}{[lgl] logical, determining if the basename should be
interpreted as filename or as last directory. If set to NULL
(default), the last entry will be interpreted if the last character is
either \ or / and as filename else.}
}
\value{
[list] with components:
\itemize{
 \item{normpath: normalized path}
 \item{drive: drive letter - only important for windows}
 \item{dirname: directory name}
 \item{fullfilename: full filename}
 \item{fullpath: full path use this - works for windows and unix}
 \item{filename: filename}
 \item{extension: extension}
}

If last.is.file is FALSE, dirname, filename, extension and fullfilename
will be NA.
}
\description{
Split a full path in its components. This is specifically an issue in
Windows and not really interesting for other OSs.
}
\examples{
vmTools:::split_path("C:/Users/JohnDoe/Documents/MyFile.txt")
}
