% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DescTools.R
\name{SplitPath}
\alias{SplitPath}
\title{Split Path In Drive, Path, Filename}
\usage{
SplitPath(path, last.is.file = NULL)
}
\arguments{
\item{path}{[chr] a path}

\item{last.is.file}{[lgl] logical, determining if the basename should be
interpreted as filename or as last directory. If set to NULL
(default), the last entry will be interpreted if the last character is
either \ or / and as filename else.}
}
\value{
[list] with components:
 - normpath: normalized path
 - drive: drive letter
 - dirname: directory name
 - fullfilename: full filename
 - fullpath: full path
 - filename: filename
 - extension: extension

 If last.is.file is FALSE, dirname, filename, extension and fullfilename
 will be NA.

 @examples
 SplitPath("C:/Users/JohnDoe/Documents/MyFile.txt")
}
\description{
replaces - DescTools::SplitPath()
}
\details{
Split a full path in its components. This is specifically an issue in
Windows and not really interesting for other OSs.
}
