% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_dir_exists}
\alias{validate_dir_exists}
\title{Validate whether a directory exists}
\usage{
validate_dir_exists(x, verbose = TRUE)
}
\arguments{
\item{x}{[path] A directory path}

\item{verbose}{[lgl] message to std_out?}
}
\value{
[lgl] TRUE if directory exists, FALSE otherwise
}
\description{
Validate whether a directory exists
}
\examples{
vmTools:::validate_dir_exists(".") # TRUE
vmTools:::validate_dir_exists("nonexistent") # FALSE
}
