% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_not_empty}
\alias{validate_not_empty}
\title{Validate an object is not length 0, empty, blank etc.}
\usage{
validate_not_empty(x)
}
\arguments{
\item{x}{[any] some argument to check}
}
\value{
[lgl] FALSE if empty in some way, TRUE otherwise
}
\description{
Designed to also catch missing args when called inside a function.
}
\seealso{
Other validations: 
\code{\link{is_an_error}()},
\code{\link{validate_dir_exists}()}
}
\concept{validations}
\keyword{internal}
