% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DescTools.R
\name{ZeroIfNA}
\alias{ZeroIfNA}
\title{Replace NAs by 0}
\usage{
ZeroIfNA(x)
}
\arguments{
\item{x}{[vector] whose NAs should be overwritten with 0s.}
}
\value{
[vector] edited vector x
}
\description{
replaces - DescTools::ZeroIfNA()
}
\details{
Replace NAs in a numeric vector x with 0. This function has the same
logic as the zeroifnull function in SQL. NAIfZero() does replace zeros
with NA. BlankIfNA() and NAIfBlank() do the same, but for character
vectors.
}
\examples{
ZeroIfNA(c(1, 2, NA, 4)) # 1 2 0 4
}
