% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_dir_exists}
\alias{assert_dir_exists}
\title{Assert a directory exists on disk}
\usage{
assert_dir_exists(x)
}
\arguments{
\item{x}{[chr] A directory path}
}
\value{
[none] stop if assertion fails
}
\description{
Assert a directory exists on disk
}
\examples{
vmTools:::assert_dir_exists(".") # OK
try(vmTools:::assert_dir_exists("nonexistent")) # Error
}
