% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_named_list}
\alias{assert_named_list}
\title{Assert an object is a list with named elements}
\usage{
assert_named_list(x)
}
\arguments{
\item{x}{[list] List to check}
}
\value{
[none] stop if assertion fails
}
\description{
Stops if:
 - x is not a list
 - x is a data.frame
 - x has no names
 - x has any NA names
 - x has any zero-length names
 - x has any whitespace-only names
}
\examples{
assert_named_list(list(a = 1, b = 2)) # OK
assert_named_list(data.frame(a = 1, b = 2)) # Error
}
