% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_scalar}
\alias{assert_scalar}
\title{Types of check functions:
1. assert_x - stop if conditions are unmet
2. validate_x - warn if conditions are unmet
              - Return TRUE/FALSE}
\usage{
assert_scalar(x)
}
\arguments{
\item{x}{[any] Element to check}
}
\value{
[none] stop if assertion fails
}
\description{
Assert an element is atomic and length 1
}
\examples{
assert_scalar("A") # OK
assert_scalar(1:2) # Error
}
