% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_scalar_not_empty}
\alias{assert_scalar_not_empty}
\title{Assert x is a scalar, and not empty in some way}
\usage{
assert_scalar_not_empty(x)
}
\arguments{
\item{x}{[any] some object to check}
}
\value{
[none] stop if assertion fails
}
\description{
Assert x is a scalar, and not empty in some way
}
\seealso{
Other assertions: 
\code{\link{assert_dir_exists}()},
\code{\link{assert_named_list}()},
\code{\link{assert_scalar}()},
\code{\link{assert_type}()},
\code{\link{assert_x_in_y}()}
}
\concept{assertions}
\keyword{internal}
