% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_scalar_not_empty}
\alias{assert_scalar_not_empty}
\title{Assert x is a scalar, and not empty in some way}
\usage{
assert_scalar_not_empty(x)
}
\arguments{
\item{x}{[any] some object to check}
}
\value{
[none] stop if assertion fails
}
\description{
Assert x is a scalar, and not empty in some way
}
\examples{
vmTools:::assert_scalar_not_empty("A") # OK
try(vmTools:::assert_scalar_not_empty(Inf)) # Error - Inf considered non-meaningful
}
\seealso{
Other assertions: 
\code{\link{assert_dir_exists}()},
\code{\link{assert_named_list}()},
\code{\link{assert_scalar}()},
\code{\link{assert_type}()}
}
\concept{assertions}
