% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_type}
\alias{assert_type}
\title{Assert an object is a scalar of a certain type}
\usage{
assert_type(x, type)
}
\arguments{
\item{x}{[any] Object to check}

\item{type}{[chr] Type to check against}
}
\value{
[none] stop if assertion fails
}
\description{
Assert an object is a scalar of a certain type
}
\seealso{
Other assertions: 
\code{\link{assert_dir_exists}()},
\code{\link{assert_named_list}()},
\code{\link{assert_scalar}()},
\code{\link{assert_scalar_not_empty}()},
\code{\link{assert_x_in_y}()}
}
\concept{assertions}
\keyword{internal}
