% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{is_an_error}
\alias{is_an_error}
\title{Determine if an object is an error}
\usage{
is_an_error(x)
}
\arguments{
\item{x}{[obj] some R object}
}
\value{
[lgl] TRUE / FALSE
}
\description{
Determine if an object is an error
}
\seealso{
Other validations: 
\code{\link{validate_dir_exists}()},
\code{\link{validate_not_empty}()}
}
\concept{validations}
