% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{is_an_error}
\alias{is_an_error}
\title{Determine if an object is an error}
\usage{
is_an_error(x)
}
\arguments{
\item{x}{[obj] some R object}
}
\value{
[lgl] TRUE / FALSE
}
\description{
Determine if an object is an error
}
\examples{
vmTools:::is_an_error(simpleError("error")) # TRUE
vmTools:::is_an_error(message("error")) # FALSE
}
