% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_depth.R
\name{lapply_depth}
\alias{lapply_depth}
\title{lapply at some list dept}
\usage{
lapply_depth(.x, .depth, .f, ...)
}
\arguments{
\item{.x}{[list] List to apply function to}

\item{.depth}{[integer] Depth to apply function at}

\item{.f}{[function] Function to apply}

\item{...}{[any] Additional arguments to pass to .f}
}
\value{
[list] List with function applied at target depth
}
\description{
Very simple replacement for purrr::map_depth to remove package dependency,
but not very robust.  Internal package use only in select cases.
}
\examples{
# intended use case - "pluck" a named list element out of each list
vmTools:::lapply_depth(list(A = list(one = 1, two = 2), B = list(one = 3, two = 4)), 1, `[[`, "one")
}
