% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_dir_exists}
\alias{validate_dir_exists}
\title{Validate whether a directory exists}
\usage{
validate_dir_exists(x, verbose = TRUE)
}
\arguments{
\item{x}{[path] A directory path}

\item{verbose}{[lgl] message to std_out?}
}
\value{
[lgl] TRUE if directory exists, FALSE otherwise
}
\description{
Validate whether a directory exists
}
\seealso{
Other validations: 
\code{\link{is_an_error}()},
\code{\link{validate_not_empty}()}
}
\concept{validations}
