% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_dir_exists}
\alias{validate_dir_exists}
\title{Validate whether a directory exists}
\usage{
validate_dir_exists(x, verbose = TRUE)
}
\description{
@param x [path] A directory path
 @param verbose [lgl] message to std_out?
}
\details{
@return [lgl] TRUE if directory exists, FALSE otherwise

 @examples
 validate_dir_exists(".") # TRUE
 validate_dir_exists("nonexistent") # FALSE
}
